skip database creation test on python 2
the database creation needs the lzma module available only in python 3.3
or later or as backport module

+++ tests/test_materials_database.py
@@ -31,31 +31,36 @@ class TestDatabase(unittest.TestCase):
     def setUpClass(cls):
         cls.el = xu.materials.elements.dummy
         # test create database
-        cls.dbfilename = tempfile.NamedTemporaryFile(delete=False).name
-        cmd = [sys.executable,
-               os.path.join(xu.__path__[0], 'materials/_create_database.py'),
-               cls.dbfilename,
-               os.path.join(os.path.dirname(__file__),
-                            '../xrayutilities/materials/data')]
-        r = subprocess.call(cmd)
-        if r != 0:
-            raise Exception('Database creation failed!')
-        cls.db = xu.materials.DataBase(cls.dbfilename)
-        cls.db.Open()
-        cls.db.SetMaterial(cls.el.name)
+        if sys.version_info >= (3, 3):
+            cls.dbfilename = tempfile.NamedTemporaryFile(delete=False).name
+            cmd = [sys.executable,
+                   os.path.join(xu.__path__[0],
+                                'materials/_create_database.py'),
+                   cls.dbfilename,
+                   os.path.join(os.path.dirname(__file__),
+                                '../xrayutilities/materials/data')]
+            r = subprocess.call(cmd)
+            if r != 0:
+                raise Exception('Database creation failed!')
+            cls.db = xu.materials.DataBase(cls.dbfilename)
+            cls.db.Open()
+            cls.db.SetMaterial(cls.el.name)
 
     @classmethod
     def tearDownClass(cls):
-        try:
-            cls.db.Close()
-            os.remove(cls.dbfilename)
-        except OSError:
-            pass
+        if sys.version_info >= (3, 3):
+            try:
+                cls.db.Close()
+                os.remove(cls.dbfilename)
+            except OSError:
+                pass
 
     def test_db_f0(self):
         f0 = self.el.f0(0)
         self.assertAlmostEqual(f0, 1.0, places=10)
 
+    @unittest.skipIf(sys.version_info < (3, 3),
+                     "needs the lzma module -> python 3.3 or newer")
     def test_owndb_f0(self):
         f0 = self.db.GetF0(0)
         self.assertAlmostEqual(f0, 1.0, places=10)
@@ -64,6 +69,8 @@ class TestDatabase(unittest.TestCase):
         f1 = self.el.f1(-1)
         self.assertTrue(math.isnan(f1))
 
+    @unittest.skipIf(sys.version_info < (3, 3),
+                     "needs the lzma module -> python 3.3 or newer")
     def test_owndb_f1_neg(self):
         f1 = self.db.GetF1(-1)
         self.assertTrue(math.isnan(f1))
@@ -72,6 +79,8 @@ class TestDatabase(unittest.TestCase):
         f1 = self.el.f1(1000)
         self.assertAlmostEqual(f1, 0.0, places=10)
 
+    @unittest.skipIf(sys.version_info < (3, 3),
+                     "needs the lzma module -> python 3.3 or newer")
     def test_owndb_f1(self):
         f1 = self.db.GetF1(1000)
         self.assertAlmostEqual(f1, 0.0, places=10)
@@ -80,6 +89,8 @@ class TestDatabase(unittest.TestCase):
         f2 = self.el.f2(-1)
         self.assertTrue(math.isnan(f2))
 
+    @unittest.skipIf(sys.version_info < (3, 3),
+                     "needs the lzma module -> python 3.3 or newer")
     def test_owndb_f2_neg(self):
         f2 = self.db.GetF2(-1)
         self.assertTrue(math.isnan(f2))
@@ -88,6 +99,8 @@ class TestDatabase(unittest.TestCase):
         f2 = self.el.f2(1000)
         self.assertAlmostEqual(f2, 0.0, places=10)
 
+    @unittest.skipIf(sys.version_info < (3, 3),
+                     "needs the lzma module -> python 3.3 or newer")
     def test_owndb_f2(self):
         f2 = self.db.GetF2(1000)
         self.assertAlmostEqual(f2, 0.0, places=10)
